Move webhookPublisher to subdir
