+++ publisher/webhookPublisher.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package fission
+package publisher
 
 import (
 	"io/ioutil"
